refactor: user 모듈의 이름을 auth로 수정
