+++ core/auth/.gitignore
