+++ core/auth/build.gradle.kts
