+++ core/auth/proguard-rules.pro
