+++ core/auth/src/main/AndroidManifest.xml
