+++ settings.gradle.kts
@@ -28,6 +28,6 @@ include(":core:firebase")
 include(":core:util")
 include(":data:domain")
 include(":data:memo")
-include(":core:user")
+include(":core:auth")
 include(":core:notification")
 include(":feature:settings")
